api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "langchain>=0.3.19",
    "langchain-azure-ai>=0.1.1",
    "langchain-mistralai>=0.2.7",
    "langchain-openai>=0.3.6",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "serpapi>=0.1.5",
    "structlog>=25.1.0",
    "uvicorn>=0.34.0",
]
